Use our own event loop
This drops the gtk event loop for our own event loop that allows us to
interleave processing of dbus events with gtk events. This allows dbus
activate events to be consumed while the window is visible, preventing
multiple activation key presses from causing the window to reappear
after it is destroyed.

This will also allow using the activate signals in the future to
refocus the window if focus has been lost.